更新sequelize
