+++ server/src/lib/sequelize.js
@@ -2,6 +2,7 @@
  * Sequelize.js api说明文档
  * https: //itbilu.com/nodejs/npm/V1PExztfb.html
  */
+// const sequelize = require('./init.sequelize')();
 const Sequelize = require('sequelize/index');
 const config = require(':config/server.base.config'); //配置文件
 const { systemLogger, sqlLog } = require(':lib/logger4');
@@ -27,31 +28,6 @@ const Attrs = (table, list) => {
     }
     return arr;
 };
-
-//多数据源 这样是没有任何提示了,可以在正式环境这样使用;
-// const dbs = Object();
-// Object.keys(config.dbs).forEach((key, index) => {
-//     const type = config.dbs[key].type;
-//     const db = config.dbs[key].config;
-//     if (type === 'sequelize') {
-//         dbs[key] = new Sequelize(db.database, db.username, db.password, {
-//             ...db.options,
-//             logging(sql) { //日志输出 不显示的输出设置为false
-//                 sqlLog.info(`${config.dbs[key].database}----${sql}`);
-//             }
-//         });
-//         dbs[key].authenticate().then((res) => {
-//             systemLogger.info(`连接数据库：${db.database} 成功!`);
-//         }).catch(err => {
-//             systemLogger.info(`连接数据库：${db.database} 出错!`, err);
-//         });
-
-//         //同步数据库模型专用 此操作将会删除数据库的表重新创建,请谨慎使用
-//         // dbs[key].sync({ force: true }).then(function(result) {
-//         //     console.log('result');
-//         // });
-//     }
-// });
 
 const BiuDB = new Sequelize(config.dbs.BiuDB.config.database, config.dbs.BiuDB.config.username, config.dbs.BiuDB.config.password, {
     ...config.dbs.BiuDB.config.options,
@@ -79,40 +55,3 @@ module.exports = {
     DataTypes: Sequelize.DataTypes,
     BiuDB
 };
-
-/**
- *
-const init = () => { //多数据源
-    const options = {
-        Sequelize,
-        SOP,
-        COL,
-        Attrs,
-        DataTypes: Sequelize.DataTypes
-    };
-    Object.keys(config.dbs).forEach((key) => {
-        const type = config.dbs[key].type;
-        const db = config.dbs[key].config;
-        if (type === 'sequelize') {
-            options['BiuDB'] = new Sequelize(db.database, db.username, db.password, {
-                ...db.options,
-                logging(sql) { //日志输出 不显示的输出设置为false
-                    sqlLog.info(`${db.database}----${sql}`);
-                }
-            });
-            options[key].authenticate().then((res) => {
-                systemLogger.info(`连接数据库：${db.database} 成功!`);
-            }).catch(err => {
-                systemLogger.info(`连接数据库：${db.database} 出错!`, err);
-            });
-
-        //同步数据库模型专用 此操作将会删除数据库的表重新创建,请谨慎使用
-        // dbs[key].sync({ force: true }).then(function(result) {
-        //     console.log('result');
-        // });
-        }
-    });
-    return options;
-};
-
- */
